fix: Do the late bindings
