+++ src/MediaTailor.ts
@@ -1,10 +1,18 @@
 import * as crypto from 'crypto';
-
+import { Lazy } from 'aws-cdk-lib';
 import {
   CfnPlaybackConfiguration,
 } from 'aws-cdk-lib/aws-mediatailor';
 
 import { Construct } from 'constructs';
+
+function removeFilename(url: string): string {
+  const lastSlash = url.lastIndexOf('/');
+  if (lastSlash === -1) {
+    return url;
+  }
+  return url.substring(0, lastSlash + 1);
+}
 
 export interface MediaTailorProps {
   readonly videoContentSourceUrl: string;
@@ -26,7 +34,11 @@ export class MediaTailor extends Construct {
     // Create EMT config
     this.config = new CfnPlaybackConfiguration(this, 'CfnPlaybackConfiguration', {
       name: `${crypto.randomUUID()}`,
-      videoContentSourceUrl,
+      videoContentSourceUrl: Lazy.string({
+        produce() {
+          return removeFilename(videoContentSourceUrl);
+        },
+      }),
       adDecisionServerUrl,
       slateAdUrl,
     });
